requirements.txt
twikit @ git+https://github.com/d60/twikit.git@main
openai>=1.0.0
python-dotenv>=1.0.0
